machined-dbus: move manager_add_machine() and _get_machine_by_pid() to -core
